+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- IOError handling for downloads and .zip extractions
 
 ## [0.4.1] - 2024-12-28
 ### Added